// src/app/@constants/enum.ts
export enum ADD_NEW_LEFT {
  ELEMENT = 'element',
  ASSETS = 'assets',
  SECTION = 'section',
  POPUP = 'popup',
  DROPBOX = 'dropbox',
  APP = 'app',
  DATASET = 'dataset',
  MEDIA = 'media',
  FOLDER = 'folder',
  FONT = 'font',
}

export enum ELEMENT_ITEM {
  HEADLINE = 'headline',
  BUTTON = 'button',
  IMAGE = 'image',
  GALLERY = 'gallery',
  BOX = 'box',
  SHAPE = 'shape',
  LINE = 'line',
  FORM = 'form',
  VIDEO = 'video',
  COLLECTION = 'collection',
  CAROUSEL = 'carousel',
  TABS = 'tabs',
  FRAME = 'frame',
  ACCORDION = 'accordion',
  TABLE = 'table',
  SURVEY = 'survey',
  HTML_CODE = 'html_code',
  PARAGRAPH = 'paragraph',
  SECTION = 'section',
}

export enum DEVICE {
  DESKTOP = 'desktop',
  MOBILE = 'mobile',
}

export enum TYPE_DRAG {
  RESIZE_SECTION = 0,
  DRAG_ELEMENT,
}

export enum RESIZE_ENUM {
  LADI_N_RESIZE = 'ladi-n-resize',
  LADI_S_RESIZE = 'ladi-s-resize',
  LADI_W_RESIZE = 'ladi-w-resize',
  LADI_E_RESIZE = 'ladi-e-resize',
  LADI_NE_RESIZE = 'ladi-ne-resize',
  LADI_NW_RESIZE = 'ladi-nw-resize',
  LADI_SE_RESIZE = 'ladi-se-resize',
  LADI_SW_RESIZE = 'ladi-sw-resize',
}